# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life Admin Companion API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface DeleteSuccess {
  success: boolean;
}

export type LifeAdminItemStatus =
  (typeof LifeAdminItemStatus)[keyof typeof LifeAdminItemStatus];

export const LifeAdminItemStatus = {
  active: "active",
  completed: "completed",
  renewed: "renewed",
  cancelled: "cancelled",
  archived: "archived",
} as const;

export type LifeAdminItemCostFrequency =
  (typeof LifeAdminItemCostFrequency)[keyof typeof LifeAdminItemCostFrequency];

export const LifeAdminItemCostFrequency = {
  "one-off": "one-off",
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
  unknown: "unknown",
} as const;

export type LifeAdminItemPriority =
  (typeof LifeAdminItemPriority)[keyof typeof LifeAdminItemPriority];

export const LifeAdminItemPriority = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface LifeAdminItem {
  id: number;
  userId: string;
  title: string;
  category: string;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  referenceNumber?: string | null;
  status: LifeAdminItemStatus;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  renewalDate?: string | null;
  /** @nullable */
  reminderDate?: string | null;
  /** @nullable */
  costAmount?: number | null;
  costFrequency: LifeAdminItemCostFrequency;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  usefulLink?: string | null;
  priority: LifeAdminItemPriority;
  createdAt: string;
  updatedAt: string;
}

export type CreateItemBodyStatus =
  (typeof CreateItemBodyStatus)[keyof typeof CreateItemBodyStatus];

export const CreateItemBodyStatus = {
  active: "active",
  completed: "completed",
  renewed: "renewed",
  cancelled: "cancelled",
  archived: "archived",
} as const;

export type CreateItemBodyCostFrequency =
  (typeof CreateItemBodyCostFrequency)[keyof typeof CreateItemBodyCostFrequency];

export const CreateItemBodyCostFrequency = {
  "one-off": "one-off",
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
  unknown: "unknown",
} as const;

export type CreateItemBodyPriority =
  (typeof CreateItemBodyPriority)[keyof typeof CreateItemBodyPriority];

export const CreateItemBodyPriority = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface CreateItemBody {
  /** @minLength 1 */
  title: string;
  category: string;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  referenceNumber?: string | null;
  status: CreateItemBodyStatus;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  renewalDate?: string | null;
  /** @nullable */
  reminderDate?: string | null;
  /** @nullable */
  costAmount?: number | null;
  costFrequency: CreateItemBodyCostFrequency;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  usefulLink?: string | null;
  priority: CreateItemBodyPriority;
}

export type UpdateItemBodyStatus =
  (typeof UpdateItemBodyStatus)[keyof typeof UpdateItemBodyStatus];

export const UpdateItemBodyStatus = {
  active: "active",
  completed: "completed",
  renewed: "renewed",
  cancelled: "cancelled",
  archived: "archived",
} as const;

export type UpdateItemBodyCostFrequency =
  (typeof UpdateItemBodyCostFrequency)[keyof typeof UpdateItemBodyCostFrequency];

export const UpdateItemBodyCostFrequency = {
  "one-off": "one-off",
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
  annually: "annually",
  unknown: "unknown",
} as const;

export type UpdateItemBodyPriority =
  (typeof UpdateItemBodyPriority)[keyof typeof UpdateItemBodyPriority];

export const UpdateItemBodyPriority = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface UpdateItemBody {
  /** @minLength 1 */
  title?: string;
  category?: string;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  referenceNumber?: string | null;
  status?: UpdateItemBodyStatus;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  renewalDate?: string | null;
  /** @nullable */
  reminderDate?: string | null;
  /** @nullable */
  costAmount?: number | null;
  costFrequency?: UpdateItemBodyCostFrequency;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  usefulLink?: string | null;
  priority?: UpdateItemBodyPriority;
}

export type DashboardSummaryByCategoryItem = {
  category: string;
  count: number;
};

export interface DashboardSummary {
  totalActive: number;
  overdueCount: number;
  dueSoonCount: number;
  dueThisMonthCount: number;
  reminderPassedCount: number;
  noDueDateCount: number;
  byCategory: DashboardSummaryByCategoryItem[];
  needsAttention: LifeAdminItem[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListItemsParams = {
  category?: string;
  status?: string;
  dueSoon?: boolean;
  search?: string;
  includeArchived?: boolean;
};

export type ListItems200 = {
  items: LifeAdminItem[];
};

export type GetUpcomingItems200 = {
  items: LifeAdminItem[];
};

export type GetRecentlyUpdatedItems200 = {
  items: LifeAdminItem[];
};