# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Life Admin Companion API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: items
    description: Life admin item operations
  - name: dashboard
    description: Dashboard summary endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List life admin items for the current user
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: dueSoon
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LifeAdminItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new life admin item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemBody'
      responses:
        '201':
          description: Created item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifeAdminItem'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get a single life admin item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifeAdminItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      operationId: updateItem
      tags: [items]
      summary: Update a life admin item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemBody'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifeAdminItem'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete a life admin item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary counts and alerts
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /dashboard/upcoming:
    get:
      operationId: getUpcomingItems
      tags: [dashboard]
      summary: Get items due in the next 30 days
      responses:
        '200':
          description: Upcoming items
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LifeAdminItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /dashboard/recent:
    get:
      operationId: getRecentlyUpdatedItems
      tags: [dashboard]
      summary: Get recently updated items
      responses:
        '200':
          description: Recently updated items
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LifeAdminItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    LifeAdminItem:
      type: object
      required:
        - id
        - userId
        - title
        - category
        - status
        - priority
        - costFrequency
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        category:
          type: string
        provider:
          type: [string, 'null']
        referenceNumber:
          type: [string, 'null']
        status:
          type: string
          enum: [active, completed, renewed, cancelled, archived]
        dueDate:
          type: [string, 'null']
          format: date
        renewalDate:
          type: [string, 'null']
          format: date
        reminderDate:
          type: [string, 'null']
          format: date
        costAmount:
          type: [number, 'null']
        costFrequency:
          type: string
          enum: [one-off, weekly, monthly, quarterly, annually, unknown]
        notes:
          type: [string, 'null']
        usefulLink:
          type: [string, 'null']
        priority:
          type: string
          enum: [low, normal, high]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateItemBody:
      type: object
      required: [title, category, status, priority, costFrequency]
      properties:
        title:
          type: string
          minLength: 1
        category:
          type: string
        provider:
          type: [string, 'null']
        referenceNumber:
          type: [string, 'null']
        status:
          type: string
          enum: [active, completed, renewed, cancelled, archived]
        dueDate:
          type: [string, 'null']
          format: date
        renewalDate:
          type: [string, 'null']
          format: date
        reminderDate:
          type: [string, 'null']
          format: date
        costAmount:
          type: [number, 'null']
        costFrequency:
          type: string
          enum: [one-off, weekly, monthly, quarterly, annually, unknown]
        notes:
          type: [string, 'null']
        usefulLink:
          type: [string, 'null']
        priority:
          type: string
          enum: [low, normal, high]

    UpdateItemBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        category:
          type: string
        provider:
          type: [string, 'null']
        referenceNumber:
          type: [string, 'null']
        status:
          type: string
          enum: [active, completed, renewed, cancelled, archived]
        dueDate:
          type: [string, 'null']
          format: date
        renewalDate:
          type: [string, 'null']
          format: date
        reminderDate:
          type: [string, 'null']
          format: date
        costAmount:
          type: [number, 'null']
        costFrequency:
          type: string
          enum: [one-off, weekly, monthly, quarterly, annually, unknown]
        notes:
          type: [string, 'null']
        usefulLink:
          type: [string, 'null']
        priority:
          type: string
          enum: [low, normal, high]

    DashboardSummary:
      type: object
      required:
        - totalActive
        - overdueCount
        - dueSoonCount
        - dueThisMonthCount
        - reminderPassedCount
        - noDueDateCount
        - byCategory
        - needsAttention
      properties:
        totalActive:
          type: integer
        overdueCount:
          type: integer
        dueSoonCount:
          type: integer
        dueThisMonthCount:
          type: integer
        reminderPassedCount:
          type: integer
        noDueDateCount:
          type: integer
        byCategory:
          type: array
          items:
            type: object
            required: [category, count]
            properties:
              category:
                type: string
              count:
                type: integer
        needsAttention:
          type: array
          items:
            $ref: '#/components/schemas/LifeAdminItem'